+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
-v0.80.4
+v0.81.0
 
-- We now always defer to the HTTP/HTTPS proxy for DNS resolution.
+- KyberPreKeyStore.markKyberPreKeyAsUsed now takes three arguments, to allow tracking how the pre-key is used: the Kyber pre-key ID, the signed EC pre-key ID, and the session base key.
+
+- We now always defer to an HTTP/HTTPS proxy for DNS resolution.